feat(friends): add own tag to friends list
